requirements.txt
torch
torchvision
streamlit
numpy
pillow
seaborn
opencv-python
matplotlib
scikit-learn